Fix RedHat 7 compiler issues and unit test issues on the same platform.
I believe this will also fix the problems on recent Ubuntu systems as well (newer kernel and related software has some file systems we can't monitor).

+++ source/code/shared/tools/scx_admin/cmdparser.h
@@ -21,7 +21,7 @@
 #include "admin_api.h"
 
 /// helper define that returns number of elements in static array
-#define DIM(x) (sizeof(x)/sizeof((x)[0]))
+#define DIM(x) (static_cast<int>(sizeof(x)/sizeof((x)[0])))
 
 
 namespace SCX_Admin {
